+++ CHANGELOG.md
@@ -8,3 +8,5 @@ Please see the separate per solution/project CHANGELOG files:
 - [JK.Common.Data.Sql](src/JK.Common.Data.Sql/CHANGELOG.md) 
 - [JK.Common.EntityFrameworkCore](src/JK.Common.EntityFrameworkCore/CHANGELOG.md) 
 - [JK.Common.EntityFrameworkCore.SqlServer](src/JK.Common.EntityFrameworkCore.SqlServer/CHANGELOG.md) 
+- [JK.Common.FluentValidation](src/JK.Common.FluentValidation/CHANGELOG.md) 
+- [JK.Common.OpenXml](src/JK.Common.OpenXml/CHANGELOG.md)